change base class dashboards
