+++ app/cells/dashboards/author.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module Dashboards
-  # This cell displays the dashboard for the mentor role
-  class Author < BaseCell
+  # This cell displays the dashboard for the author role
+  class Author < BaseDashboard
   end
 end
