+++ app/cells/dashboards/member.rb
@@ -2,6 +2,6 @@
 
 module Dashboards
   # This cell displays the dashboard for the member role
-  class Member < BaseCell
+  class Member < BaseDashboard
   end
 end
